backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pillow==10.4.0
pandas>=2.1.0
python-dotenv==1.0.1
openai==1.43.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
supabase==2.3.0
aiofiles==23.2.1
